requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
astroid==3.0.3
attrs==25.3.0
black==23.12.0
click==8.1.8
dill==0.4.0
frozenlist==1.6.0
grpcio==1.60.0
grpcio-tools==1.60.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
multidict==6.4.3
mypy-extensions==1.0.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
propcache==0.3.1
protobuf==4.25.1
pylint==3.0.3
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==79.0.0
six==1.17.0
tomlkit==0.13.2
tzdata==2025.2
yarl==1.20.0
